requirements.txt
Django==1.6.2
South==0.8.4
argparse==1.2.1
distribute==0.6.24
django-appconf==0.6
django-compressor==1.3
django-endless-pagination==2.0
ipdb==0.8
ipython==2.0.0
oauthlib==0.6.1
phishtank==1.0.1
requests==2.1.0
requests-oauthlib==0.4.0
six==1.6.1
twython==3.1.2
wsgiref==0.1.2
